pyproject.toml
[project]
name = "sifthub-assignment"
version = "1.0.0"
description = "An assignment"
authors = []
readme = "README.md"
requires-python = ">=3.10,<3.14"
dependencies = [
    "fastapi (>=0.118.3,<0.119.0)",
    "uvicorn (>=0.37.0,<0.38.0)",
    "python-dotenv (>=1.1.1,<2.0.0)",
    "python-multipart (>=0.0.20,<0.0.21)",
    "tiktoken (>=0.12.0,<0.13.0)",
    "langchain (>=0.3.27,<0.4.0)",
    "pymupdf (>=1.26.5,<2.0.0)",
    "chromadb (>=1.1.1,<2.0.0)",
    "cohere (>=5.18.0,<6.0.0)",
    "voyageai (>=0.3.5,<0.4.0)",
    "ranx (>=0.3.21,<0.4.0)",
    "numpy (==2.2.6)"
]

[tool.poetry]
package-mode = false

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
